# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is a sample server Content server.",
    "version": "1.0.0",
    "title": "Swagger Content",
    "termsOfService": "http://swagger.io/terms/",
    "contact": { "email": "apiteam@swagger.io" },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "basePath": "/v1",
  "tags": [
    {
      "name": "post",
      "description": "Everything about your Posts",
      "externalDocs": {
        "description": "Find out more",
        "url": "http://swagger.io"
      }
    }
  ],
  "schemes": ["http", "https"],
  "paths": {
    "/posts": {
      "post": {
        "tags": ["post"],
        "summary": "Add a new post",
        "description": "",
        "operationId": "addPost",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "the object of post",
            "required": true,
            "schema": { "$ref": "#/definitions/post" }
          }
        ],
        "responses": { "405": { "description": "Invalid input" } }
      },
      "get": {
        "tags": ["post"],
        "summary": "List posts",
        "description": "List all posts",
        "operationId": "listPosts",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "keyword",
            "in": "query",
            "description": "Keyword for filter",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": { "$ref": "#/definitions/post" }
            }
          },
          "400": { "description": "Invalid status value" }
        }
      }
    },
    "/posts/{postId}": {
      "get": {
        "tags": ["post"],
        "summary": "Find post by ID",
        "description": "Returns a single post",
        "operationId": "getPostById",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "description": "ID of post to return",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": { "$ref": "#/definitions/post" }
          },
          "400": { "description": "Invalid ID supplied" },
          "404": { "description": "post not found" }
        }
      },
      "put": {
        "tags": ["post"],
        "summary": "Update a post",
        "description": "Update a post",
        "operationId": "updatePost",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "description": "ID of post",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "in": "body",
            "name": "body",
            "description": "the object of post",
            "required": true,
            "schema": { "$ref": "#/definitions/post" }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": { "$ref": "#/definitions/post" }
          },
          "405": { "description": "Invalid input" }
        }
      },
      "delete": {
        "tags": ["post"],
        "summary": "Deletes a post",
        "description": "",
        "operationId": "deletePost",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "description": "post id to delete",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "400": { "description": "Invalid ID supplied" },
          "404": { "description": "post not found" }
        }
      }
    }
  },
  "securityDefinitions": {
    "consumer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "definitions": {
    "post": {
      "type": "object",
      "required": ["name", "photoUrls"],
      "properties": {
        "id": { "type": "integer", "format": "int64" },
        "title": { "type": "string", "example": "doggie" },
        "content": {
          "type": "string",
          "example": "blabla  blabal \n balbalbalalb \n bababab"
        },
        "photoUrls": {
          "type": "array",
          "xml": { "name": "photoUrl", "wrapped": true },
          "items": { "type": "string" }
        },
        "tags": {
          "type": "array",
          "items": { "type": "string" }
        }
      }
    },
    "ApiErrorResponse": {
      "type": "object",
      "properties": {
        "code": { "type": "integer", "format": "int32" },
        "type": { "type": "string" },
        "message": { "type": "string" }
      }
    }
  },
  "externalDocs": {
    "description": "Find out more about Swagger",
    "url": "http://swagger.io"
  }
}